Don't start audio service in foreground (#1122)
Since the audio service is always started when the app is in the
foreground, and since there are still many ANRs due to not calling
startForeground quickly enough (though it's the earliest thing that
happens in the service), there is likely a race condition in which at
certain points of time, audio playback is stopped and removes the
notification and stops the service before startForeground is called. As
a workaround, just start the service in the background. This is possible
since the app is in the foreground at this point in time.