Implement a way to expand tafseer links (#1129)
In some tafaseer, a single "entry" gives the translations for multiple
ayat. Previously, the app would just show a "same as translation for
ayah X." This made many people sad. This patch attemps to fix this by
adding the ability to tap the translation to actually display the
translation text.