Fix override parameter conflict
Because QuranScreenInfo is no longer a singleton (intentionally due to
the fact that the page size calculator changes between page types), it
is possible to have 2 QuranScreenInfo instances around at the same time,
one pointing to an override parameter and the other not pointing to one.

This particular solution isn't efficient (requiring a SharedPrefs lookup
before either of the get width methods are called), but it's the
quickest and safest solution for the time being.