Refactor some of QuranDataActivity's logic (#1155)
This patch moves a bunch of code from QuranDataActivty out into a
QuranDataPresenter. It also sends a reconnection broadcast upon resuming
(so that if no events have been emitted and a download is in progress,
the activity can know about it). It also caches the last successful
checking of pages (where all the pages are present and no patch is
necessary) to avoid running the checks again.

Note that a full caching would be problematic in the case that the
person didn't have the files, started downloading, and, upon coming
back, the app tried to download again (even though it was already
done downloading while in the background).