Re-add the partial page checker
This patch re-adds the partial page checker. It also fixes some bugs in
the previous code for it (such as the 1260 images always getting a
handful of pages deleted by the checker whenever it runs, thus forcing
the download twice). This fixes #1252, though at some point, this code
should really be removed completely.