Log some additional events (#1262)
Log when the partial page checker removes images. Also, write a file in
the internal storage directory to try to figure out if internal files
are also being deleted when images are lost (if the theory is correct,
the times when images are lost and this file is lost should be few,
representing only the times when the app had its data cleared, was
uninstalled and reinstalled, etc. If there are cases in which the
internal file survives but the sdcard files don't, this is good proof of
an sdcard cleaning up removing this data.