Fix images not loading when in airplane mode
This fixes a bad bug introduced in #1247. The code added a check for
internet before attempting to load the image (thinking it was checking
before attempting to download the image). This moves the internet
checking code to the right place and renames the method to avoid
confusion in the future.