Downloading image should take a width parameter
The per page downloader attempts to download a page. If it fails due to
out of memory, for example, it attempts to fallback to the tablet page
width (if the device is in dual page mode). Because the downloader
always tries to use the current width, this parameter is unfortunately
ignored, and the app tries to redownload the same page width again.