Enable multidex
Sadly need to enable multidex. This is to be able to add WorkManager in
a future PR (but even if not, the app is so close to the dex limit
anyway, that it's just a matter of time until it is necessary to
multidex). For now, this likely only affects debug builds, since after
r8, the app is likely still single dex.