Fix bugs with the partial page checker and tablet (#1276)
Al7amdulillah, got really lucky here! The parital page checker had a
number of bugs - namely it used the width where it should have used the
tablet width. Consequently, it marks a massive number of pages as "to be
deleted." Due to another bug though (looping over only the pages to
delete instead of all pages to delete), al7amdulillah the pages didn't
actually get deleted in this case. This fixes these bugs.