Integrate with the audio update api (#1318)
This patch implements a weekly check for audio updates. Upon finding
updates, it figures out which files actually require updating and
removes them. A notification is sent to the person to inform them that
some files were erased. An improved ui can be implemented at some point
in the future.