Fix for a case where start is larger than end
We have a check to make sure we are not trying to play "backwards" - for
some reason, in some rare cases, the end ayah is less than the start
ayah. This checks and silently handles the case, while logging it as a
checked exception so that the root cause can be fixed.