Target api 29 and migrate for scoped storage
In order to target api 29 (which is required for apps to update as of
November 2020), prepare to support scoped storage. This patch changes
the default directory to the application storage directory for the
application. It also attempts to migrate people on apis 29 and above who
are still using the sdcard to the application storage directory. This
should make it easy to target api 30 when necessary. Fixes #1246.