Move towards ayah menu consolidation (#1457)
This patch consolidates a lot of the ayah menu handling for the
translation view with that used in the Quran views. This allows for
getting close to feature parity with the long press menu of the Quran
screen itself. Consequently, this allows bookmarking and tagging
directly from the translation view.

Also made the selection behavior closer to that of the main page, in
that if you have an ayah selected and click another one, the selection
moves to that ayah. Unlike the main view, however, translation view
still doesn't support multiple ayah selection.

Fixes #220.