Put page in its place in split translation view (#1461)
In split screen between Quran and translation, leave the Quran page
where it is in the normal mushaf instead of having a preference to pin
it to the left or right. This way, pages match their expected place in
the mushaf instead of being pinned to one side or the other.