Fix repeat sometimes stopping at end of file (#1509)
Gapless files sometimes have a 999 entry, and sometimes don't. When they
don't, playback continues until the end of the file. If a file doesn't
have a 999 entry, it's possible for the file to end before a seek
happens, and for the seek to happen when we're trying to replay the file
insted of seek. This causes an error and causes the playback to stop.

This patch works around this by properly seeking if the end of file is
reached in a gapless file (instead of trying to re-play).