Allow upgrade logic per page type
Today, there's a lot of code within the initial data check and upgrade
path that's specific to a certain page type - for example, the database
is bundled for madani databases, the pages may be bundled in the future
for warsh databases, and so on.

This patch adds an interface wherein different page types can intercept
and handle updates on their own before going to the server for data.