Fallback to the internal directory when pages lost (#1568)
Despite changing the directory to the internal directory, there's still
a case where people on the old external directory get their files
deleted. This will prompt the app to ask for downloading the files,
which will again download in the same place, causing the files to be
lost again due to the app cleaner. This patch causes the location of the
data to switch to the internal storage when pages are detected to have
been lost.