Allow updating bookmarks during prefs update
This patch updates preference updates to be able to update and change
bookmarks. This is relevant in warsh, for example, where the app
switches from kufi counting to madani counting and needs to update
bookmarks accordingly. Using this, a flavor can have logic run on update
of preferences which modifies bookmarks.