Support new page types and fix some bugs
This patch adds a field to PageProvider to handle different types of
pages - for example, in the future, pages could be rendered per line, or
using fonts, etc. It also increases the maximum number of files allowed
in a zip file limit.

This also fixes a bug that causes the page selection to crash after
scrolling to the very end.