Rename QuranPageModule to QuranDataModule
There are 2 QuranPageModules today, one that has the data and is part of
the singleton scope of the app, and the other is one that's really per
page view. This is confusing, so renamed one to QuranDataModule since
it's really about Quran data. A few fields in there probably make sure
to move at some point to QuranPageModule or similar.