Clean up ayah selections logic
Previously, the ayah selection logic was heavily intertwined with
PagerActivity, where some state was owned by the PagerActivity and other
state was owned by the fragments. Moreover, PagerActivity had to go in
to the ViewPager, look for every fragment, cast it, and ask it to handle
things to handle ayah selection.

This is the first patch to clean this up and makes ayah selection
reactive instead.