Fix scope on ReadingEventPresenter
In a previous patch, ReadingEventPresenter was changed from activity
scope to singleton scope, mainly because some of the fragments in the
bottom panel were injected with an application scope. This fixes it by
injecting those with an activity scope.

Tags is a bit of a special case since it is shared both by the bottom
panel, and as a dialog in other cases. This bottom panel version needs
some extra things (observing the stream of changes, for example) that
the dialog version doesn't. This patch thus also splits the tag dialog
to have a subclass just for the bottom panel, and moves the relevant
logic to it.