Fallback to images if page types fail to download
Whereas the normal full pages can be loaded on demand, the new line by
line images need to all be present for them to be used. This patch
updates the image checking code to fall back to the full madani pages if
the line by line pages fail to download completely.

Refs #1589.