Respect first ayah start time
Today, when starting at the first ayah, we'd return the position of the
"0th" ayah to get a basmallah, etc if present. This causes an issue when
there is no "0th" ayah set. In those cases, the file starts playback at
the very start instead of at the position where the first ayah should
start. This patch fixes this issue.