Minor performance improvements
onCreateOptionsMenu is re-called every time the menu is invalidated in a
toolbar world. This can be expensive, especially when we are inflating a
SearchView each time. This patch fixes this by caching the menu item and
modifying the enabled state of it instead.

Moreover, replace a weight 1 on a single view with match parent to avoid
unnecessary measurement of spinner children.