Migrate bookmarks and recents on page type change
This patch migrates bookmarks bookmarks and recent pages when the page
type changes to pages with a different number of overall pages. Note
that this is imprecise for page bookmarks, since the first ayah on a
page in the madani mushaf may be the last ayah on a page in shemerly or
vice versa.