Fix audio highlight transition crashes
In some cases, audio playback would fail once animation was reenabled
due to the conversion of the code from the TypeEvaluator and related
classes from Java to Kotlin. During this conversion, the lists were
copied instead of being allowed to mutate, causing normalization of the
arrays to the same size to fail, thus causing index out of bounds
exceptions.