Allow removing a certain page from bookmarks
In some cases, when we swap out pages, we might get rid of extra pages
that are no longer there. This allows removing bookmarks for such pages.
This is really meant for pages like 612 (the du3a page) in Naskh, in
preparation for replacing this set with the King Fahd images, which do
not have this page.