Clean up hacks for Shemerly in tablet
Before this patch, tablet mode (displaying two pages side by side in
landscape) would always count the second page as the page (for all
purposes - that of the title, bookmarks, recent pages, etc). This fact,
coupled with the fact that Shemerly has an odd number of pages, caused
some issues on the very last page of Shemerly.

Hacks were added in various places to subtract one from pages to support
Shemerly properly, especially with respect to the last page. This patch
cleans all this up by instead updating the tablet logic to consider the
first of the two pages as the primary page for all purposes (title,
bookmarks, recent pages, etc).