Audio position update time should factor in speed
The audio service continuously sends events checking the playback status
every few seconds, depending on how much time is left until the end of
the ayah. These update timings need to now factor in the speed, so if
we are 3 seconds away from the end of the ayah and are playing at 1.5x,
for example, checking after 3 seconds is too late (since we would have
played some of the next ayah).