Fix audio not starting after required downloads
Certain files, when downloaded, were not notifying the app that the
download is complete (due to an ordering bug of the "file download"
being sent after the "batch download done," when the batch should be
last). This caused required downloads to not continue to playing audio
when downloaded. This patch fixes this.