Batch database with ordered audio download batches
Today, the download of consecutive suras happens as a "batch," and the
database is a separate "batch," breaking the dialog and what not. This
fixes it by placing the database download within the consecutive suras
batch.

For non-consecutive suras, each sura is a download "batch," and the
database is also a separate "batch" today. This will change in a future
PR in sha' Allah.