Refactor some of QuranDownloadService
Replace code within the service with code in a separate class, split by
the type. This will make it easier to add "batches" to replace the
current way download manager downloads non-ranged downloads in the
future. It can also enable some testing of this logic in the future
as well in sha' Allah.