Add BatchDownloadStrategy
This is a not very efficient download strategy to download a batch of
non-consecutive suras. It does many more allocations compared to the
consecutive downloaders, but is still significantly more efficient than
the previous version of n disparate download requests. In a future
commit, this will be improved in sha' Allah.