Split Madani module into 3 parts
This PR splits Madani into 3 modules instead of one - instead of just
pages:madani, there's now a pages:madani, which depends on a
pages:common:madani and a pages:data:madani. The data module now
contains the data source. The common module includes things like upgrade
and size calculators. The pages:madani remains the main Madani module,
depending on the other two.

This patch also opens the code for the upgrades, which was previously
closed.