Make some improvements to Arabic search
This patch makes some improvements to Arabic search. It matches some
additional matches for alif and hamza, and, more importantly, it removes
the limit for suggestions (it used to only search the first 250 ayat
during suggestions mode) in order to give people results that are more
likely to be expected.