Fix incorrect toolbar position
This patch fixes a regression introduced during the migration to support
edge to edge in 8d7f9f749d34c078ff77af080e3bdcc55a09919a. This isn't an
ideal fix, but it's the same as the code has been before. In the future,
we should remove the forced LTR on the root RelativeLayout and the
corresponding switch back to RTL of the toolbar.

Note that the marker isn't ideal for some ayat, but this was also the
case before the update, so it should be fixed separately in sha' Allah.

Fixes #2985.