Fix page bookmarks not working in some cases
This patch fixes page bookmarks. Due to recent refactoring to make the
bookmarks observable, bookmarking a page and then swiping a few pages in
either direction and coming back would cause the bookmark to flicker and
then come back. This is because the logic to refresh the bookmark status
in onPrepareOptionsMenu was removed during the refactoring. This patch
adds it back.