Remove right to left hack on root layout
Previously, in order to ensure the ayah toolbar was set properly, a hack
was in place to always ensure that the root RelativeLayout is LTR.
Removing this hack broke the ayah toolbar position. Adding it back,
however, broke the proper directionality of the audio bar.

This patch finally removes this hack and properly calculates the
position in RTL. Consequently, it fixes the audio bar directionality.