Fix many more toolbar issues
After the last PR to fix some RTL toolbar issues, there were other
issues that were missed. For example, in landscape RTL, the toolbar was
broken. Moreover, it was broken for translations. This patch ensures
that the toolbar is well behaved in RTL and LTR in portrait and
landscape and in translation mode and normal mode (including on tablet
with dual screen and split screen modes).

The remaining issues not addressed by this PR are:

1. in landscape, when the toolbar is visible in translation mode and a
long press happens on an ayah, the y value is incorrectly lower than
the ayah box.

2. in tablet split screen mode, long pressing a translation and then
scrolling the translation down does not move the toolbar down.

Refs #2993.