Fix the position of the page number for line pages
This patch fixes a bug that caused the page number to be on the wrong
side while in Arabic. This is likely due to removing the hack that had
the parent layout for the entire page as an LTR layout. Consequently,
when this flag was removed, the actual page numbers were sometimes
rendered the opposite direction due to smart row components flipping the
direction of the children for RTL.