Use dark mode as the default
For now, set the default mode to be dark mode. In the future, in sha'
Allah, there will be a setting for either day, night, or follow system.
For people who have the app installed, it will default to dark. For new
installations, it'll follow system by default. This will be in the
future though, in sha' Allah.