Fix audio settings button missing on some devices
This patch fixes the audio settings button going missing on some devices
(devices with width of 320dp, for example). It sets a width per button
to ensure that everything fits, and assigns weight of 1 to each so that
the buttons can have more space if the space available is more.