Fix wrong page when switching foldable state
When switching between folded and open states, the page would often
reset (ex from sura Nur ayah 35 in closed mode would become sura Nisa'
in open mode). This is because there are two conversions happening. By
setting isDualPages, the currentPage read is the new position, but the
code would then re-convert that page, thus breaking it. This updates it
by reading currentPage before updating isDualPages.

In the future, this code can likely be simplified (i.e. removing the
mapping in favor of just reading the most up to date page).