Fix audio playback bugs due to setting speed
This patch fixes bugs due to setting the playback speed on certain
devices, even if the person does not try to change the playback speed at
all. For example, on a Pixel 8, pausing the audio and resuming causes
the playback to stop ui wise without stopping service wise due to a
silent exception from setting the playback params.

The fix here, as suggested by Claude, is not re-using the playback
params, and, instead, setting a new one, irrespective of the current
values.