Fix translations sometimes being hidden
This patch fixes an issue where the translations in the inline
translations view shows no visible translations. This happens because,
at some point, all translations were disabled (or the list of enabled
translations is out of sync with the actual available translations).

The patch fixes this by enabling all available translations when no
available translations are selected.

Fixes #3144.