Add support for opus files
This adds support for different audio formats and places the logic to
decide which format to use behind an interface so that it can be tweaked
later if necessary.

For now, all new qaris will default to opus only. Based on this, can
decide what to do on existing qaris that get opus support in the future
in sha' Allah.