Fix icon colors on some versions of Android
The QuranRowFactory was getting tint colors using the application
context, which does not know about the current theming situation. This
updates it so that a resource id is sent instead, and then is resolved
using the correct activity context.

Ideally, nothing in QuranRowFactory should need a context, but
unfortunately, strings do at the moment, so this will have to be cleaned
up sometime in the future to avoid it causing issues when swapping
between Arabic and non-Arabic locales.