Use ExoPlayer instead of MediaPlayer
ExoPlayer is much more resilient and customizable and is the recommended
route today. For Quran specifically, it has improved seeking, which
stops vbr audio seeks from being extremely inaccurate, and stops us from
having to force encode everything as cbr.

It also handles things for us (like audio focus and ducking) resulting
in less code to maintain. Since it is in use for Android Auto also,
there is no additional apk size to pay on top of today's apk size.