Let ExoPlayer seek to the correct position
Previously, with MediaPlayer, a file would be prepared for playback and
then seek would be called when it's ready to correctly go to the correct
position. ExoPlayer allows passing in a parameter of the time at which
to begin the playback.

Aside from cleaning the code, this fixes a bug where, after the
migration to ExoPlayer, if a sura is playing and an attempt is made to
play an ayah in the middle of another sura, it would start the other
sura without seeking to the correct ayah, instead starting at the
beginning.