Do not wait for all DNS resolvers to return
Previously, due to the coroutine scope, MultiDns was forced to await all
dns providers to finish before returning. Since the lookups are blocking
calls and do not respect cancellation, this could take some time. This
patch fixes it so they're in a separate coroutine scope, so that we
don't have to wait on the cancellation before returning.