Fix missing basmallah for gapless playback
When playing from a gapless sura, the MediaPlayer code used to start at
the timestamp for ayah=0 (usually the basmallah for suras that aren't
Fatiha nor Tawba), falling back to ayah=1 otherwise. During the
migration to ExoPlayer, this was inadvertently changed to always start
at ayah=1, preventing the basmallah from playing.