Fix basmallah between repeats
Due to a migration bug when moving from MediaPlayer to ExoPlayer, the
timing updater was run while the basmallah was playing in between sura
transitions during repeat ranges, causing the basmallah to not play or
be cut off. This fixes it to not restart the timing updater for player
overrides (basmallah playbacks).